Work on the animations API

+++ src/server/ws/dto/StatusDto.h
@@ -15,19 +15,16 @@ namespace creatures :: ws {
         DTO_FIELD_INFO(status) {
             info->description = "Short status text";
         }
-
         DTO_FIELD(String, status);
 
         DTO_FIELD_INFO(code) {
             info->description = "Status code";
         }
-
-        DTO_FIELD(Int32, code);
+        DTO_FIELD(UInt16, code);
 
         DTO_FIELD_INFO(message) {
             info->description = "Verbose message";
         }
-
         DTO_FIELD(String, message);
 
     };
